chore(app): retire the 426 update-floor client, keep the build-identity header (PRODUCT-1144)
The gateway's min-app-version floor is gone (cloud PR #234), so the desktop
half of that contract comes out: the 426 forwarder in both gateway
transports, the __HOUSTON_UPDATE_REQUIRED__ global, the update-required
bus + /v1/version minAppVersion probe, the blocking UpdateRequired screen
(+ its en/es/pt strings and analytics events), and staging.sh's floor
pre-flight warning.

X-Houston-App-Version STAYS: update-floor.ts slims down to app-version.ts
(channel + header bake) in both app and web-adapter halves, and every
transport still attaches the header. Nothing server-side acts on it now,
but it keeps gateway logs attributable to a build, and every target's CORS
allow-set already admits it — dropping it would only buy another
coordinated header/CORS dance (HOU-825) to ever bring back.

usage.gen.json regenerated (UpdateRequired specimen gone). The engine
client's Version type drops minAppVersion — the gateway never emits it
anymore and its only readers were the deleted floor probe.

typecheck, biome lint/format, boundaries, and all suites pass except
ui/engine-client's two store-catalog fixture tests, which fail identically
on clean origin/main (skills: [] drift — pre-existing).
