+++ app/src/locales/es/shell.json
@@ -393,14 +393,6 @@
     "reassurance": "Tus agentes siguen trabajando en segundo plano y tus chats y configuraciones quedan guardados. Vas a retomar justo donde lo dejaste.",
     "updateNowAction": "Actualizar ahora"
   },
-  "updateRequired": {
-    "cardLabel": "Actualización de Houston requerida",
-    "title": "Actualización requerida",
-    "description": "Se necesita Houston {{minVersion}} o más reciente para seguir usando Houston Cloud. Estás en v{{currentVersion}}.",
-    "descriptionNoVersion": "Se necesita una versión más reciente de Houston para seguir usando Houston Cloud. Estás en v{{currentVersion}}.",
-    "downloadAction": "Descargar la actualización",
-    "checkAction": "Buscar actualizaciones"
-  },
   "naming": {
     "dialogTitle": "Nombra tu agente",
     "newAgentFallback": "Nuevo agente",
