+++ packages/host/src/server.ts
@@ -208,9 +208,9 @@ export interface ControlPlaneDeps {
 
 function applyCors(deps: ControlPlaneDeps, res: ServerResponse): void {
   res.setHeader("Access-Control-Allow-Origin", deps.corsOrigin || "*");
-  // X-Houston-App-Version: the desktop app's update-floor identity header
-  // (app-update floor) — sent by its shared gateway transport to this host
-  // too, so the preflight must allow it even though the host ignores it.
+  // X-Houston-App-Version: the desktop app's build-identity header — sent by
+  // its shared gateway transport to this host too, so the preflight must
+  // allow it even though the host ignores it.
   res.setHeader(
     "Access-Control-Allow-Headers",
     "Authorization, Content-Type, X-Houston-App-Version",
